Expand null check methods used in PreconditionsCheckNotNullPrimitive and merge it in UnnecessaryCheckNotNull
RELNOTES: Expand null check methods used in PreconditionsCheckNotNullPrimitive and merge it in UnnecessaryCheckNotNull

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=256251942
