+++ core/src/main/java/com/google/errorprone/scanner/BuiltInCheckerSuppliers.java
@@ -218,7 +218,6 @@ import com.google.errorprone.bugpatterns.PackageInfo;
 import com.google.errorprone.bugpatterns.PackageLocation;
 import com.google.errorprone.bugpatterns.ParameterComment;
 import com.google.errorprone.bugpatterns.ParameterName;
-import com.google.errorprone.bugpatterns.PreconditionsCheckNotNullPrimitive;
 import com.google.errorprone.bugpatterns.PreconditionsCheckNotNullRepeated;
 import com.google.errorprone.bugpatterns.PreconditionsInvalidPlaceholder;
 import com.google.errorprone.bugpatterns.PredicateIncompatibleType;
@@ -563,7 +562,6 @@ public class BuiltInCheckerSuppliers {
           PeriodFrom.class,
           PeriodTimeMath.class,
           PeriodGetTemporalUnit.class,
-          PreconditionsCheckNotNullPrimitive.class,
           PredicateIncompatibleType.class,
           PrivateSecurityContractProtoAccess.class,
           ProtocolBufferOrdinal.class,
