+++ docs/bugpattern/nullness/UnnecessaryCheckNotNull.md
@@ -15,3 +15,11 @@ still checked to be null and is thus redundant.
 *   Often the order of the two arguments is swapped, and the reference is never
     actually checked for nullity. This check ensures that the first argument to
     such methods is not a literal.
+
+*   When a primitive is passed as the argument to check, the primitive will be
+    [autoboxed](http://docs.oracle.com/javase/7/docs/technotes/guides/language/autoboxing.html)
+    into a boxed object, which is non-null, causing the check to always pass
+    without the condition being evaluated. If the intent was to ensure that the
+    primitive met some criterion (e.g., a boolean that should be non-null),
+    please use `Preconditions.checkState()` or `Preconditions.checkArgument()`
+    instead.
